+++ lib/nugrant.rb
@@ -1,6 +1,5 @@
 require 'nugrant'
 require 'nugrant/config'
-require 'nugrant/parameter_bag'
 require 'nugrant/parameters'
 
 unless defined?(KeyError)